python_service/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy>=2.0.36

# HTTP client
httpx==0.26.0

# AI / ML - Zero-Shot Classification
transformers>=4.30.0
torch>=2.0.0

# Embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0

# OpenAI-compatible client for vLLM
openai>=1.0.0

# Image processing (optional, for image validation)
pillow>=10.0.0

# OCR (Tesseract)
pytesseract>=0.3.10
opencv-python-headless>=4.8.0

# Constraint Solving (Z3 SMT Solver)
z3-solver>=4.12.0

# OpenTelemetry - Distributed Tracing
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-sqlalchemy>=0.45b0
opentelemetry-instrumentation-logging>=0.45b0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Linting
ruff==0.1.14
